# Conflicts:
#	server/src/service/statistic/statistic.repository.implementation.ts

+++ server/src/service/room/room.service.ts
@@ -158,7 +158,7 @@ export class RoomService {
     });
 
     const operation$ = Promise.resolve(true);
-    if (token) {
+    if (token && room) {
       this.playersBindService.removePlayers(room.token, token);
     }
